% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Averages photographs of the same type}
\usage{
aggregateData(projectName, replicate = c("lines", "type"), varFunc = "se")
}
\arguments{
\item{projectName}{the short name in use for the project.}

\item{replicate}{which columns indicate the factors to use}

\item{varFunc}{what type of variation measurment to perform. Currently supports \code{varFunc} = "se" to calculate the standard error, \code{varFun} = "cv" to calculate the coefficient of variation or any built-in R function (e.g., sd).}
}
\value{
A dataframe "projectName.ag" is saved to the global environment and a .csv file "projectName_ag.csv" is exported to the "parameter_files" directory.
}
\description{
Addin description here
}
\author{
Aleeza C. Gerstein
}

