% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoCrop.R
\name{AutoCrop}
\alias{AutoCrop}
\title{Run Fiji/ImageJ auto-crop macro on a folder of images}
\usage{
AutoCrop(
  photoDir,
  outputDir,
  imageJLoc = NA,
  macroName = "auto_crop.ijm",
  debug = FALSE
)
}
\arguments{
\item{photoDir}{Path to the input image folder.}

\item{outputDir}{Path to the output folder where cropped images will be saved.}

\item{imageJLoc}{Absolute path to Fiji/ImageJ app folder. If \code{NA},
the function will try common default locations on macOS.}

\item{macroName}{Name of the macro file inside the installed package.
Defaults to \code{"auto_crop.ijm"}.}

\item{debug}{Logical. Whether to print debug messages.}
}
\value{
Invisibly returns the normalized output directory path.
}
\description{
\code{AutoCrop} runs a Fiji/ImageJ macro stored in the package
to automatically crop 90mm plate images from a black background.
}
