% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotRaw.R
\name{discplotNoRep}
\alias{discplotNoRep}
\title{Used to plot the results of a single picture analyzed in imageJ}
\usage{
discplotNoRep(data, label = label, ymin = 0, ymax = 250, xmin = 0,
  xmax = 40, standardLoc = 2.5, cexPt = 0.6, stand = 0, xaxt = "n",
  yaxt = "n", plotStandardLoc = FALSE)
}
\arguments{
\item{data}{a dataframe containing two columns, with the x coordinates and image intensity height. Typically stored in a list named from the projectName in \code{\link{runIJ}}, \code{\link{runIJManual}} or \code{\link{readInExisting}}}

\item{label}{what to title the plot}

\item{ymin}{a numeric value indicating the minimum y value plotted in each graph}

\item{ymax}{a numeric value indicating the maximum y value plotted in each graph}

\item{xmin}{a numeric value indicating the minimum x value plotted in each graph}

\item{xmax}{a numeric value indicating the maximum x value plotted in each graph}

\item{standardLoc}{a numeric value indicating the location on the disk in mm used to standardize pixel intensity across all pictures, defaults to 2.5}

\item{cexPt}{a numeric value indicating the size to plot for points, x-axis text and y-axis text, respectively}

\item{xaxt}{a character string to determine whether to plot the x axis, defaults to "n"}

\item{yaxt}{a character string to determine whether to plot the y axis, defaults to "n"}

\item{plotStandardLoc}{a logical value indicating whether to draw a dashed horizontal line at the standardization point}
}
\description{
This function is used by \code{\link{plotRaw}} to plot the results from each photograph
}

