% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{find.mle}
\alias{find.mle}
\title{Helper fuctions from diversitree package}
\usage{
find.mle(func, x.init, method, ...)
}
\arguments{
\item{func}{interal useage}

\item{x.init}{interal useage}

\item{method}{interal useage}

\item{...}{interal useage}
}
\description{
\code{mle}  These functions are from the package diversitree which is no longer able to be loaded from CRAN or github. They are used by maxLik
}
