% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageJInterface.R
\docType{class}
\name{imageJInterface-class}
\alias{imageJInterface-class}
\alias{imageJInterface}
\title{Creates a a thin wrapper reference class for imageJ
Automatically verifies the locaiton of the imageJ binaries on the user's system}
\arguments{
\item{filePath}{Allows the full file path to the imageJ application to be set (optional).}
}
\value{
A reference class object
}
\description{
Creates a a thin wrapper reference class for imageJ
Automatically verifies the locaiton of the imageJ binaries on the user's system
}

\author{
Erik White \email{erikwhite@gmail.com}
}
