% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhibGrow.R
\name{inhibGrow}
\alias{inhibGrow}
\title{Find the points of inhibition, maximum growth and resource limitations}
\usage{
inhibGrow(projectName, diskDiam = 12.7, maxDist = 30)
}
\arguments{
\item{projectName}{the short name in use for the project.}

\item{diskDiam}{the diameter of the diffusion disk in mm, defaults to 6.}

\item{maxDist}{a numeric value indicating the maximum distance away from the disk to be considered. Defaults to 30mm.}

\item{nameVector}{either a logial value indicating whether to plot the photograph names above the graph or not or a vector the same length as the number of pictures containing the desired names. Defaults to TRUE.}
}
\value{
A dataframe "projectName_inhib.df" is saved to the global environment and a .csv file "projectName_inhib_df.csv" is exported to the "parameter_files" directory.
}
\description{
\code{inhibGrow}  uses the intensity information to find the length of inhibition, points of maximum growth, and points of resource limitation.
}
\details{
\code{\link{inhibGrow}} is useful for compounds that both inhibit growth yet are also utilized. In comparison to \code{\link{maxLik}}, the growth profile for these compounds don't reach an asymptote, but rather are characterized by inhibition, followed by a point of maximum growth and then a growth reduction.
The pixel intensity information previously determined from \code{\link{IJMacro}} is used to determine the growth maximum and inhibition points.
Slopes are calculated from the minimum intensity to the point of maximum intensity ('slope2Max') and the point of maximum intensity to the edge of the plate ('slopeFromMax').
}
\examples{
\dontrun{
inhibGrow("myProject")
maxLik("myProject", diskDiam = 12, maxDist = 25)
}
}
