% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/minorFunctions.R
\name{is.letter}
\alias{is.letter}
\title{Used to determine whether something is a letter}
\usage{
is.letter(x)
}
\arguments{
\item{x}{a single element}
}
\description{
Used to determine whether something is a letter
}

