% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotRaw.R
\name{plotRaw}
\alias{plotRaw}
\title{Used to plot the results of the imageJ analysis}
\usage{
plotRaw(projectName, ymin = 0, ymax = 250, xmin = 0, xmax = 40,
  xplots = 6, height = 4, width = 8, cexPt = 0.6, cexX = 0.8,
  cexY = 0.8, nameVector = TRUE, plotStandardLoc = TRUE,
  showNum = FALSE, popUp = TRUE, overwrite = TRUE)
}
\arguments{
\item{projectName}{the short name to be used for the project}

\item{ymin}{a numeric value indicating the minimum y value plotted in each graph}

\item{ymax}{a numeric value indicating the maximum y value plotted in each graph}

\item{xmin}{a numeric value indicating the minimum x value plotted in each graph}

\item{xmax}{a numeric value indicating the maximum x value plotted in each graph}

\item{xplots}{a numeric value indicating how many plots to plot in each row}

\item{height}{a numeric value indicating the height of the pdf file generated}

\item{width}{a numeric value indicating the width of the pdf file generated}

\item{cexPt,}{cexX, cexY a numeric value indicating the size to plot for points, x-axis text and y-axis text, respectively}

\item{nameVector}{either a logial value indicating whether to plot the name of the picture above the graph or not or a vector the same length as the number of pictures containing the desired names. Defaults to TRUE.}

\item{plotStandardLoc}{a logical value indicating whether to draw a dashed horizontal line at the standardization point}

\item{showNum}{a logical value indicating whether to annotate each graph with the picture number (determined alphabetically from the picture names)}

\item{popUp}{a logical value indicating whether to pop up the figure after it has been created}

\item{overwrite}{a logical value indicating whether to overwrite existing figures created on the same day for the same project name}
}
\value{
A pdf file with one plot for each photograph is saved to visualize the results of imageJ analyses
}
\description{
This function creates a pdf figure of plots showing the results of the imageJ analysis, with one plot for each picture. This function is primarily for visualization purposes.
}

