% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readExistingDF.R
\name{readExistingDF}
\alias{readExistingDF}
\title{Read in an existing dataframe using the tcltk interface}
\usage{
readExistingDF(projectName)
}
\arguments{
\item{projectName}{the short name you want use for the project. Note, this can be different than the previous file name used for this project.}
}
\value{
\code{projectName} is saved to the global directory and can be used for \code{aggregateData}
}
\description{
Open an existing dataframe previously created with either \code{createDataframe} or \code{aggregateData} using tcltk interface.
}
\author{
Aleeza C. Gerstein
}
\seealso{
\code{\link{runIJManual}} to run the imageJ analysis by manually supplying the project folder and photo location paths
}

