% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readInExistingIJ.R
\name{readInExisting}
\alias{readInExisting}
\title{Used to read in existing imageJ analyses}
\usage{
readInExisting(projectName, newList = list(), numDig = 30)
}
\arguments{
\item{projectName}{the name to be used for project. This name should be short, and can be different than what was used originally for the imageJ analysis step.}
}
\description{
This function can be used to read in existing imageJ analyses following either \code{\link{runIJ}} or \code{\link{runIJManual}}. Running this function will prompt the user to select the main project folder and to select the directory that contains
}

