% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readExistingIJ.R
\name{readInExistingIJ}
\alias{readInExistingIJ}
\title{Used to read in existing imageJ analyses}
\usage{
readInExistingIJ(projectName, newList = list())
}
\arguments{
\item{projectName}{the name to be used for project. This name should be short, and can be different than what was used originally for the imageJ analysis step.}

\item{newList}{dummy variable}
}
\description{
This function can be used to read in existing imageJ analyses following \code{\link{IJMacro}}. Running this function will prompt the user to select the main project folder and to select the directory that contains
}

