% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/runIJ.R
\name{runIJ}
\alias{runIJ}
\title{Run the imageJ analysis using the tcltk interface}
\usage{
runIJ(projectName, imageJLoc = "default", diskDiam = 6)
}
\arguments{
\item{projectName}{the short name you want use for the project.}

\item{imageJLoc}{the absolute path to imageJ on your computer. Current options are those standard for a mac: \code{imageJLoc} = "default" when imageJ is located at /Applications/ImageJ/ImageJ.app/Contents/MacOS/JavaApplicationStub; \code{imageJLoc} = "loc2" for path /Applications/ImageJ.app/Contents/MacOS/JavaApplicationStub; \code{imageJ} will also accept any absolute path}

\item{diskDiam}{the diameter of the diffusion disk in mm, defaults to 6.}
}
\value{
A .csv file is saved to the directory "imageJ-out" in the directory specified by \code{projectDir}. The average line for each photograph is saved to the list \code{projectName} in the global environment.
}
\description{
Used to run the imageJ analysis component of diskImageR using tcltk interface (i.e., pop-up boxes) to identify the project directory (the directory where all subsequent analyses will be saved) as well as the location of the directory containing the photos to be analyzed.
}
\details{
Each photograph in the directory specified by \code{pictureDir} is input into ImageJ, where the built-in ‘find particles’ macro is used to find the center of a drug diffusion disk of the size specified by \code{diskDiam}. Lines are drawn every 5° out from the center of the disk, and the pixel intensity, which corresponds to cell density, is measured using the ‘plot-profile’ macro along each line. The results from all lines are saved into the "imageJ-out" directory in the specified \code{projectDir}. The average pixel intensity is then determined across all 72 lines at each distance and saved to \code{projectName}. \cr Note that the photograph names can be fairly important downstream and should follow a fairly strict convention to be able to take advantage of some of the built-in functions. Photographs should be named "line_factor1_factor2_factor3_...".
}
\author{
Aleeza C. Gerstein
}
\seealso{
\code{\link{runIJManual}} to run the imageJ analysis by manually supplying the project folder and photo location paths
}

