% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveMLparam.R
\name{saveMLparam}
\alias{saveMLparam}
\title{Save maximum likelihood output}
\usage{
saveMLparam(projectName, typical = TRUE)
}
\arguments{
\item{projectName}{the short name in use for the project.}

\item{typical}{if TRUE, a logistic curve will be calculated for each photo. If FALSE, the function will determine whether logistic, paradoxical, or confounding is the best fit and then calculate the curve.}
}
\value{
A dataframe "projectName_ML.df" and a dataframe "projectName_ML2.df"are saved to the global environment and two .csv files "projectName_ML.csv" and "projectName_ML2.csv" are exported to the "parameter_files" directory when typical=TRUE, otherwise this function returns a dataframe "projectName_ML2.df" in global environment and a .csv file "projectName_ML2.csv" in "parameter_files" directory.
}
\description{
Saves the output of maximum likelihood functions - asym, od50, scal, sigma, lnLik and type.
}
\author{
Aleeza C. Gerstein
}
