% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/addType.R
\name{addType}
\alias{addType}
\title{Add factor column}
\usage{
addType(projectName, typePlace = 3, typeName = "type2")
}
\arguments{
\item{projectName}{the short name in use for the project.}

\item{typePlace}{a number that indicates the position of the photograph name to be stored as the 'type' vector'. Defaults to 3. For more details see \code{\link{IJMacro}}}

\item{typeName}{a character string that indicates what to name the typeVector. Defaults to "type2".}
}
\value{
updates the existing dataframe
}
\description{
Add an extra factor ("type") column to the existing dataframe
}
\author{
Aleeza C. Gerstein
}

