% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcMIC.R
\name{calcMIC}
\alias{calcMIC}
\title{Convert RAD to MIC based on built-in or provided parameters or datasets}
\usage{
calcMIC(projectName, type = "df", RAD = "20")
}
\arguments{
\item{projectName}{the short name you have been using for the project.}

\item{type}{specify whether the dataset to use is a dataframe with all data ("df") or an aggregated dataframe ("ag")}

\item{RAD}{a numeric value the the critical level of the radius of inhibition (i.e., resistance) parameter to use for MIC. Currently only \code{RAD} = "80" (80\% reduction in growth), \code{RAD} = "50" (50\% reduction in growth), and \code{RAD} = "20" (20\% reduction in growth) are supported [Default = "20".}
}
\value{
An updated .csv file that contains the MIC values that correspond to calculated RAD values is saved to the directory "parameter_files" in the main project directory. The average line for each photograph is saved to the list \code{projectName} in the global environment.
}
\description{
Used to convert RAD into MIC based on a standard curve.
}
\author{
Aleeza C. Gerstein
}

