% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/createDataframe.R
\name{createDataframe}
\alias{createDataframe}
\title{Dataframe creation}
\usage{
createDataframe(projectName, clearHalo, diskDiam = 6, maxDist = 30,
  nameVector = TRUE, typeVector = TRUE, typePlace = 2,
  typeName = "type")
}
\arguments{
\item{projectName}{the short name in use for the project.}

\item{clearHalo}{numeric value that indicates which picture should be used to represent a clear halo (i.e., the clear space beside the disk). This is extremely important to determine tolerance, as the intensity beside the disk for the photograph chosen is subtracted for all photographs. Choosing the correct photograph to be use is the only subjective aspect of this pipeline; lighting and camera settings will determine the degre to which the hue of the plate backbground changes among different photographs. Care should be taken to ensure that the plate background will be as similar as possible between plates, and thus any plate with a clear area beside the disk can be chosen. Photos are numbered alphabetically by name, and can also be found using \code{\link{plotRaw}}, showNum = TRUE.}

\item{diskDiam}{the diameter of the diffusion disk in mm, defaults to 6.}

\item{maxDist}{a numeric value indicating the maximum distance away from the disk to be considered. Defaults to 30mm.}

\item{nameVector}{either a logial value or a character vector. Supported values are \code{nameVector} = "TRUE" to assign the photograph name to the 'name' column, \code{nameVector} = "FALSE" to assign th photograph number to the 'name' column, or \code{nameVector} = a vector the same length as the number of photographs indicating the desired names.}

\item{typeVector}{a logical value. \code{typeVector} = "TRUE" will add a 'type' vector to the dataframe using values found in the \code{typePlace} position of the photograph names (see \code{\link{IJMacro}} for more details) while \code{typeVector} = "FALSE" will not add a type column.}

\item{typePlace}{a number that indicates the position of the photograph name to be stored as the 'type' vector'. Defaults to 2. For more details see \code{\link{IJMacro}}}

\item{typeName}{a character string that indicates what to name the typeVector. Defaults to "type".}
}
\value{
A dataframe "projectName.df" is saved to the global environment and a .csv file "projectName_df.csv" is exported to the "parameter_files" directory.
}
\description{
Writes the main dataframe with resistance, tolerance and sensitivity parameters
}
\details{
A dataframe with 11 columns:
\itemize{
		\item\bold{name:} determined by \code{nameVector}, either photograph names, photograph numbers, or a user-supplied list of names
	 	\item\bold{line:} the first components of the \code{namesVector}; everything that comes before the first "_" in the photograph name
		\item\bold{type:} the location within the \code{name} of the phograph type is supplied by \code{typePlace}. Use \code{\link{addType}} if more than one type column are desired.
		\item\bold{ZOI80, ZOI50, ZOI20:} resistance parameters, coresponding to the distance in mm of 80\%, 50\% and 20\% reduction in growth
		\item\bold{fAUC80, fAUC50, fAUC20:} tolerance parameters, coresponding to the fraction of growth achieved above the 80\%, 50\% and 20\% reduction in growth points
		\item\bold{slope:} sensitivity parameter, indicating the slope at the midpoint, i.e., how rapidly the population changes from low growth to full growth
	}
}
\author{
Aleeza C. Gerstein
}

