% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAPArea.R
\name{PAPArea}
\alias{PAPArea}
\title{Run Fiji/ImageJ PAP area quantification macro on a folder of images}
\usage{
PAPArea(inputDir, projectDir, roiZip = NULL, imageJLoc = NA, debug = FALSE)
}
\arguments{
\item{inputDir}{Path to the input image folder.}

\item{projectDir}{Path to the output folder where result files will be saved.}

\item{roiZip}{Path to the ROI zip file. If \code{NULL}, the function will try
to use the default ROI template \code{papROISet.zip} stored in the installed package.}

\item{imageJLoc}{Absolute path to Fiji/ImageJ app folder. If \code{NA},
the function will try common default locations on macOS.}

\item{debug}{Logical. Whether to print debug messages.}
}
\value{
Invisibly returns the normalized output directory path.
}
\description{
\code{PAPArea} runs a Fiji/ImageJ macro stored in the package
to quantify PAP positive area percentage for 6-well plate images
using one common threshold per image.
}
