% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readExistingDF.R
\name{readExistingDF}
\alias{readExistingDF}
\title{Read in an existing dataframe using the tcltk interface}
\usage{
readExistingDF(projectName)
}
\arguments{
\item{projectName}{is the short name you want use for the project. Note, this can be different than the previous \code{projectName} used.}
}
\value{
\code{projectName} is saved to the global directory and can be used for \code{aggregateData}.
}
\description{
Open an existing dataframe previously created with either \code{\link{createDataframe}} or \code{\link{aggregateData}} using tcltk interface.
}
\author{
Aleeza C. Gerstein
}

