% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveMLParam.R
\name{saveMLParam}
\alias{saveMLParam}
\title{Save maximum likelihood output}
\usage{
saveMLParam(projectName)
}
\arguments{
\item{projectName}{the short name in use for the project.}
}
\value{
A dataframe "projectName_ML.df" is saved to the global environment and a .csv file "projectName_ML.csv" is exported to the "parameter_files" directory.
}
\description{
Saves the output of maximum likelihood functions - asym, od50, scal, sigma and lnLik.
}
\author{
Aleeza C. Gerstein
}

